pyproject.toml
[project]
name = "app"
version = "0.1.1"
description = "Tech Challenge Fase 1 - Warley Lima"
authors = [
    {name = "Warley Lima"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pytest (>=8.3.5,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^1.0.0"

